requirements.txt
en-core-web-sm
faiss-cpu==1.8.0
huggingface_hub
langchain
langchain-community
langchain-core
langchain-openai==0.0.7
Markdown==3.6
nltk
numpy==1.26.4
openai==1.12.0
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==15.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
PyMuPDF==1.23.26
PyMuPDFb==1.23.22
pypdf
python-dateutil
PyYAML
requests
rpds-py==0.18.0
rsa==4.9
safetensors
scikit-learn
scipy
sentence-transformers
sentencepiece
setuptools==69.1.1
spacy 
spacy-legacy
spacy-loggers
streamlit==1.33.0
tiktoken==0.6.0
tokenizers
toml==0.10.2
toolz==0.12.1
torch 
torchvision
tornado==6.4
tqdm
transformers
types-pytz
typing-inspect
typing_extensions 
uritemplate==4.1.1
urllib3
validators==0.22.0
weaviate-client==4.5.1
wheel==0.42.0